Диплом/backup/diplom1/requirements.txt
appdirs==1.4.3
asgiref==3.3.4
astroid==2.4.2
certifi==2020.11.8
cffi==1.14.5
chardet==3.0.4
click==7.1.2
colorama==0.4.3
cryptography==3.4.7
distlib==0.3.0
Django==3.2
django-filter==2.4.0
djangorestframework==3.12.4
filelock==3.0.12
Flask==1.1.2
idna==2.10
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.3
jwt==1.2.0
keyboard==0.13.5
lazy-object-proxy==1.4.3
Markdown==3.3.4
MarkupSafe==1.1.1
mccabe==0.6.1
MouseInfo==0.1.3
numpy==1.19.3
Pillow==8.0.0
PyAutoGUI==0.9.52
pycparser==2.20
pygame==2.0.0
PyGetWindow==0.0.9
pylint==2.5.3
PyMsgBox==1.0.9
PyMySQL==1.0.2
pynput==1.7.3
pyperclip==1.8.1
PyQt5==5.15.1
PyQt5-sip==12.8.1
pyqt5-tools==5.15.1.1.7.5
PyRect==0.1.4
PyScreeze==0.1.26
python-dotenv==0.14.0
PyTweening==1.0.3
pytz==2020.1
requests==2.25.0
six==1.14.0
sqlparse==0.3.1
tk==0.1.0
toml==0.10.1
urllib3==1.26.2
virtualenv==20.0.20
Werkzeug==1.0.1
wrapt==1.12.1
